+++ Modules/Intent.Modules.Application.MediatR.Behaviours/Templates/TemplateExtensions.cs
@@ -2,7 +2,6 @@ using System.Collections.Generic;
 using Intent.Modules.Application.MediatR.Behaviours.Templates.AuthorizationBehaviour;
 using Intent.Modules.Application.MediatR.Behaviours.Templates.EventBusPublishBehaviour;
 using Intent.Modules.Application.MediatR.Behaviours.Templates.LoggingBehaviour;
-using Intent.Modules.Application.MediatR.Behaviours.Templates.MongoDbUnitOfWorkBehaviour;
 using Intent.Modules.Application.MediatR.Behaviours.Templates.PerformanceBehaviour;
 using Intent.Modules.Application.MediatR.Behaviours.Templates.UnhandledExceptionBehaviour;
 using Intent.Modules.Application.MediatR.Behaviours.Templates.UnitOfWorkBehaviour;
@@ -29,11 +28,6 @@ namespace Intent.Modules.Application.MediatR.Behaviours.Templates
         public static string GetLoggingBehaviourName<T>(this IntentTemplateBase<T> template)
         {
             return template.GetTypeName(LoggingBehaviourTemplate.TemplateId);
-        }
-
-        public static string GetMongoDbUnitOfWorkBehaviourName<T>(this IntentTemplateBase<T> template)
-        {
-            return template.GetTypeName(MongoDbUnitOfWorkBehaviourTemplate.TemplateId);
         }
 
         public static string GetPerformanceBehaviourName<T>(this IntentTemplateBase<T> template)