+++ Modules/Intent.Modules.UnitOfWork.Persistence.Shared/TemplatIds.cs
@@ -1,0 +1,8 @@
+namespace Intent.Modules.UnitOfWork.Persistence.Shared;
+
+internal static class TemplateIds
+{
+    public const string DaprStateStoreUnitOfWorkInterface = "Intent.Dapr.AspNetCore.StateManagement.DaprStateStoreUnitOfWorkInterface";
+    public const string CosmosDBUnitOfWorkInterface = "Intent.CosmosDB.CosmosDBUnitOfWorkInterface";
+    public const string MongoDbUnitOfWorkInterface = "Intent.MongoDb.MongoDbUnitOfWorkInterface";
+}