+++ Modules/Intent.Modules.Application.MediatR.Behaviours/release-notes.md
@@ -1,3 +1,10 @@
+### Version 4.2.2
+
+- Improvement: `UnitOfWorkBehaviour` will now universally save changes for all of the following modules without separate `Behaviour`s being created for each:
+  - `Intent.CosmosDB`
+  - `Intent.Dapr.AspNetCore.StateManagement`
+  - `Intent.MongoDb`
+
 ### Version 4.1.3
 
 - Upgrade *Breaking Changes*: Upgraded to MediatR v12, which has breaking changes from.